requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
twilio>=8.0.0
